removed models/ folder and restructured what was inside it
