+++ deeprankcore/features/exposure.py
@@ -4,9 +4,9 @@ from Bio.PDB.PDBParser import PDBParser
 from Bio.PDB.ResidueDepth import get_surface, residue_depth
 from Bio.PDB.HSExposure import HSExposureCA
 from deeprankcore.domain import nodefeatures
-from deeprankcore.models.structure.atom import Atom
-from deeprankcore.models.structure.residue import Residue
-from deeprankcore.models.graph import Graph
+from deeprankcore.molstruct.atom import Atom
+from deeprankcore.molstruct.residue import Residue
+from deeprankcore.operations.graph import Graph
 
 
 logging.getLogger(__name__)
