+++ deeprankcore/molstruct/residue.py
@@ -1,8 +1,8 @@
 from typing import Optional
 import numpy as np
-from deeprankcore.models.structure.pdb_structure import Chain
-from deeprankcore.models.structure.aminoacid import AminoAcid
-from deeprankcore.models.pssm import PssmRow
+from deeprankcore.molstruct.pdb_structure import Chain
+from deeprankcore.molstruct.aminoacid import AminoAcid
+from deeprankcore.operations.pssm import PssmRow
 
 
 class Residue:
