+++ deeprankcore/operations/parsers/residue.py
