+++ deeprankcore/operations/pdb.py
@@ -4,10 +4,10 @@ import numpy as np
 import subprocess
 from scipy.spatial import distance_matrix
 from pdb2sql import interface as get_interface
-from deeprankcore.models.structure.pdb_structure import Chain, Structure
-from deeprankcore.models.structure.atom import Atom, AtomicElement
-from deeprankcore.models.structure.residue import Residue
-from deeprankcore.models.contact import Pair
+from deeprankcore.molstruct.pdb_structure import Chain, Structure
+from deeprankcore.molstruct.atom import Atom, AtomicElement
+from deeprankcore.molstruct.residue import Residue
+from deeprankcore.molstruct.contact import Pair
 from deeprankcore.domain.aminoacids import amino_acids
 
 _log = logging.getLogger(__name__)
