+++ tests/features/test_aminoacid.py
@@ -1,9 +1,9 @@
 from pdb2sql import pdb2sql
 from deeprankcore.domain.aminoacids import serine
-from deeprankcore.models.structure.variant import SingleResidueVariant
+from deeprankcore.molstruct.variant import SingleResidueVariant
 from deeprankcore.features.surfacearea import add_features
-from deeprankcore.models.graph import build_residue_graph
-from deeprankcore.models.operations.pdb import get_structure, get_surrounding_residues
+from deeprankcore.operations.graph import build_residue_graph
+from deeprankcore.operations.pdb import get_structure, get_surrounding_residues
 from deeprankcore.domain import nodefeatures
 from deeprankcore.features.aminoacid import add_features # noqa
