+++ tests/features/test_surfacearea.py
@@ -1,12 +1,12 @@
 from pdb2sql import pdb2sql
 import numpy as np
 from deeprankcore.domain.aminoacids import alanine
-from deeprankcore.models.structure.pdb_structure import Structure, Chain
-from deeprankcore.models.structure.residue import Residue
-from deeprankcore.models.structure.variant import SingleResidueVariant
+from deeprankcore.molstruct.pdb_structure import Structure, Chain
+from deeprankcore.molstruct.residue import Residue
+from deeprankcore.molstruct.variant import SingleResidueVariant
 from deeprankcore.features.surfacearea import add_features
-from deeprankcore.models.graph import build_residue_graph, build_atomic_graph
-from deeprankcore.models.operations.pdb import (
+from deeprankcore.operations.graph import build_residue_graph, build_atomic_graph
+from deeprankcore.operations.pdb import (
     get_structure,
     get_residue_contact_pairs,
     get_surrounding_residues)
