+++ tests/models/test_graph.py
@@ -4,10 +4,10 @@ import os
 import h5py
 from pdb2sql import pdb2sql
 import numpy as np
-from deeprankcore.models.grid import GridSettings, MapMethod
-from deeprankcore.models.graph import Graph, Edge, Node
-from deeprankcore.models.contact import ResidueContact
-from deeprankcore.models.operations.pdb import get_structure
+from deeprankcore.operations.grid import GridSettings, MapMethod
+from deeprankcore.operations.graph import Graph, Edge, Node
+from deeprankcore.molstruct.contact import ResidueContact
+from deeprankcore.operations.pdb import get_structure
 from deeprankcore.domain import (nodefeatures as Nfeat,
                                 edgefeatures as Efeat)
